refactor(entity/dto): agregando validation en el backend para la entidad y dto
